RAG/requirements.txt
openai
docling
pydantic
qdrant-client
neo4j
python-dotenv
fastapi
uvicorn
